# Speed up extraction of global props, headlines, and title