--- a/tests/test-org-roam.el
+++ b/tests/test-org-roam.el
@@ -255,7 +255,7 @@
     (it "extracts headlines"
       (expect (test #'org-roam--extract-headlines
                     "headlines/headline.org")
-              :to-equal
+              :to-have-same-items-as
               `(["e84d0630-efad-4017-9059-5ef917908823" ,(test-org-roam--abs-path "headlines/headline.org")]
                 ["801b58eb-97e2-435f-a33e-ff59a2f0c213" ,(test-org-roam--abs-path "headlines/headline.org")])))))