###### Motivation for this change

Org-roam takes a long time to process files. In addition, big files (10000+ lines) can take several seconds to process after it is modified. This becomes unbearable for eg. habit (`org-habit`) tracking files.

This PR introduces some optimizations:

- Replace unnecessary uses of `org-element-parse-buffer`
  - `org-roam--extract-headlines`: Use `org-map-region` and `org-entry-get` instead of `org-element-parse-buffer`.
  - `org-roam--extract-titles-headline`: Find the first headline with `org-outline-regexp-bol` instead of using `org-element-parse-buffer`.
- If it's available, use `org-collect-keywords` which is a lot faster as it also avoids `org-element-parse-buffer`. (As it is a relatively new function in Org, [added in April 2020](https://code.orgmode.org/bzg/org-mode/commit/b4e91b7e944c900db6b8217d78011afcd2c1e62c), I have kept the previous implementation if it's not present.)

## Performance comparison

`vocab.org` is an Org file with 14795 lines (1052 headings).

I'm running this on a Ryzen 1400 CPU.

```elisp
(with-current-buffer "vocab.org"
  (list
   (list 'total 'gc-count 'gc-time)
   (benchmark-run 1
     (org-roam--extract-headlines))
   (benchmark-run 1
     (org-roam--extract-titles-title))
   (benchmark-run 1
     (org-roam--extract-titles-headline))
   (benchmark-run 1
     (org-roam--extract-titles-alias))))
```

Before:

| total       | gc-count | gc-time            |
|-------------|----------|--------------------|
| 1.771265479 | 1        | 0.7655653740000048 |
| 0.932756371 | 0        | 0.0                |
| 1.645399532 | 1        | 0.7050031620000254 |
| 1.701283592 | 1        | 0.742148186999998  |

After:

| total       | gc-count | gc-time |
|-------------|----------|---------|
| 0.149613758 | 0        | 0.0     |
| 0.00223786  | 0        | 0.0     |
| 8.2539e-05  | 0        | 0.0     |
| 0.001961503 | 0        | 0.0     |

From 5 seconds down to 0.15 seconds.