@jethrokuan 2020-08-24
The changes LGTM.

It looks from the tests like the headline extraction now returns the candidates in reverse order, so the tests also need to be fixed here.