@kisaragi-hiu 2020-08-24
> It looks from the tests like the headline extraction now returns the candidates in reverse order, so the tests also need to be fixed here.

Ah, yes, because `push` adds to the front. I've made it so that the test uses `:to-have-same-items-as` to ignore the ordering of its items.